requirements.txt
streamlit
pandas 
google-generativeai 
matplotlib 
seaborn 
wordcloud 
openpyxl 
scikit-learn 
scipy